requirements.txt
pony==0.7.16
uvicorn==0.19.0
fastapi==0.85.0
pydantic==1.10.2
email-validator==1.3.0
pytest==7.1.3
requests==2.27.1
PyJWT==2.5.0
coverage==6.5.0
python-multipart==0.0.5
websockets==10.4
nest-asyncio==1.5.6
numpy==1.22.3
timeout-decorator==0.5.0
